backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
faster-whisper
pydub
pydantic
python-dotenv
aiofiles
